aura_agent_app/api_server/requirements.txt

fastapi>=0.135.3
uvicorn>=0.44.0
sqlalchemy>=2.0.49
psycopg2-binary>=2.9.11
pydantic>=2.12.5
pydantic-settings>=2.12.0
pydantic[email]>=2.12.0
redis>=7.4.0
langchain>=1.2.15
langchain-community>=0.4.1
langgraph>=1.1.6
langchain-openai>=1.1.13
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pycryptodome>=3.23.0
